organized the code and made general improvements
